fix: serializer
